sample/requirements.txt
flask
python-jose[cryptography]
requests
PyJWT
gunicorn
authlib